# swagger.json
{
   "swagger": "2.0",
   "info": {
     "title": "API CHATBOT MSG",
     "description": "base_url: http://51.38.234.54/api-chat",
     "version": "1.0"
   },
   "produces": ["application/json"],
  "consumes": ["application/json"],
  "paths": {
     "/conversation": {
       "post": {
         "x-swagger-router-controller": "bar",
         "operationId": "1",
         "description": "Créer une conversation",
         "parameters": [{
           "name": "body",
           "in": "body",
           "required": true,
           "schema": {
             "$ref": "#/definitions/conversation"
           }
         }
         ],
         "responses": {
           "200" : {
             "description": "Success"
           }
         }
       },
       "delete": {
         "x-swagger-router-controller": "bar",
         "operationId": "3",
         "description": "Supprimer une conversation",
         "parameters": [{
           "name": "body",
           "in": "body",
           "required": true,
           "schema": {
             "$ref": "#/definitions/conversationDelete"
           }
         }
         ],
         "responses": {
           "200" : {
             "description": "Success"
           }
         }
       }
     },
     "/conversation/{uid}" :{
       "get": {
         "x-swagger-router-controller": "bar",
         "operationId": "2",
         "description": "Récupérer une conversation",
         "parameters": [{
           "name": "uid",
           "type" : "string",
           "in": "path",
           "required": true
         }
         ],
         "responses": {
           "200" : {
             "description": "Success"
           }
         }
       }
     },
     "/message": {
		"post": {
		  "x-swagger-router-controller": "bar",
		  "operationId": "4",
		  "description": "Envoyer un message",
		  "responses": {},
		  "parameters": [{
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/message"
            },
            "name": "body"
          }
          ]
		}
	  },
      "/retrieveMessages": {
        "post": {
          "x-swagger-router-controller": "bar",
          "operationId": "5",
          "description": "Récupérer les messages non lus d'une conversation",
          "responses": {},
          "parameters": [{
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/retrieveMessages"
            },
            "name": "body"
          }
          ]
        }
      },
      "/updateMessages": {
        "post": {
          "x-swagger-router-controller": "bar",
          "operationId": "6",
          "description": "Mettre à jour les messages non lus d'une conversation",
          "responses": {},
          "parameters": [{
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/updateMessages"
            },
            "name": "body"
          }
          ]
        }
      }
   },
  "definitions": {
    "conversation" : {
      "properties" : {
        "token" : {
          "type" : "string"
        },
        "members" : {
          "type" : "array",
          "items" : {
            "type" : "string"
          }
        },
        "conversation" : {
          "type" : "object",
          "allOf": [
            {
              "$ref": "#/definitions/conversationObj"
            }
          ]
        }
      }
    },
    "conversationDelete" : {
      "properties" : {
        "token" : {
          "type" : "string"
        },
        "uid" : {
          "type" : "string"
        }
      }
    },
    "message" : {
      "properties" : {
        "uid" : {
          "type" : "string"
        },
        "message" : {
          "type" : "object",
          "allOf": [
            {
              "$ref": "#/definitions/messageObj"
            }
          ]
        }
      }
    },
    "retrieveMessages" : {
      "properties" : {
        "uid" : {
          "type" : "string"
        },
        "member" : {
          "type" : "string"
        }
      }
    },
    "updateMessages" : {
      "properties" : {
        "uid" : {
          "type" : "string"
        },
        "member" : {
          "type" : "string"
        }
      }
    },
    "conversationObj": {
      "type": "object",
      "required": [
        "uid"
      ],
      "properties": {
        "uid": {
          "type": "string"
        }
      }
    },
    "messageObj": {
      "type": "object",
      "required": [
        "uid"
      ],
      "properties": {
        "type": {
          "type": "string"
        },
        "member": {
          "type": "string"
        },
        "text" : {
          "type": "string"
        },
        "date" : {
          "type": "string"
        }
      }
    }
  }
}